+++ bitcointx/core/sha256.py
@@ -70,6 +70,10 @@ T_CSHA256 = TypeVar('T_CSHA256', bound='CSHA256')
 class CSHA256():
     __slots__ = ['s', 'buf', 'bytes_count']
 
+    buf: bytes
+    bytes_count: int
+    s: List[int]
+
     # Initialize SHA-256 state.
     def __init__(self) -> None:
         self.Reset()
@@ -240,7 +244,7 @@ class CSHA256():
             data = data[remainder_len:]
             self.bytes_count += remainder_len
             self.Transform(buf, 1)
-            self.buf: bytes = b''
+            self.buf = b''
             bufsize = 0
 
         if len(data) >= 64:
@@ -285,14 +289,14 @@ class CSHA256():
     def Reset(self) -> 'CSHA256':
         self.buf = b''  # type: bytes
         self.bytes_count = 0  # type: int
-        self.s: List[int] = [0x6a09e667,
-                             0xbb67ae85,
-                             0x3c6ef372,
-                             0xa54ff53a,
-                             0x510e527f,
-                             0x9b05688c,
-                             0x1f83d9ab,
-                             0x5be0cd19]
+        self.s = [0x6a09e667,
+                  0xbb67ae85,
+                  0x3c6ef372,
+                  0xa54ff53a,
+                  0x510e527f,
+                  0x9b05688c,
+                  0x1f83d9ab,
+                  0x5be0cd19]
         return self